feat: initial price_calculator
This is the first factor of the pricing calculator.
It handles input validation, suggestions for input on mistyped item codes and applying discounts.

Opportunities for improvement (refactor):
There are three classes being consumed by the frontend (console), a single PriceCalculator class could be using these classes as implementation details. The inventory and record managment could  be behind a repository, model abstraction.